UPDATE: Added State Transitions, Text and More
So i had some trouble getting this to run on my computer at home, for some reason it didnt want to read the SDL files, so i had to copy paste to a project that i had running before to work on it there. Then i copy and pasted it back to the original one so i can post it to git. I made a new branch just in case it might break stuff.

I added a text component and tried to do click buttons, but ive unfortunatly run out of time and opted to just use the Enter Key to transition between the states. MainMenu and EndGame both will transition to GamePlay and I have it so it doenst work on gameplay. I also Have "This is Brick Breaker", and "Would you Like to tray again?" as titles on the different states.